requirements.txt
# chromadb>=1.5.0
# google-generativeai>=0.8.6
# langchain>=1.2.10
# langchain-community>=0.4.1
# langchain-huggingface>=1.2.0
# langchain-text-splitters>=1.1.0
# pathlib>=1.0.1
# pypdf>=6.7.0
# # torch==2.2.0+cpu
# # sentence-transformers>=2.2.2
# streamlit

streamlit
chromadb
python-dotenv
google-generativeai
requests
